Application Source code/requirements.txt
altgraph==0.17.4
annotated-types==0.6.0
asttokens==2.4.1
blis==0.7.11
bottle==0.12.25
bottle-websocket==0.2.9
catalogue==2.0.10
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.16.0
colorama==0.4.6
comm==0.2.2
confection==0.1.4
contourpy==1.2.0
customtkinter==5.2.2
cycler==0.12.1
cymem==2.0.8
darkdetect==0.8.0
debugpy==1.8.1
decorator==5.1.1
Eel==0.16.0
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl#sha256=86cc141f63942d4b2c5fcee06630fd6f904788d2f0ab005cce45aadb8fb73889
et-xmlfile==1.1.0
executing==2.0.1
fasttext-wheel==0.9.2
fonttools==4.50.0
future==1.0.0
gevent==24.2.1
gevent-websocket==0.10.1
greenlet==3.0.3
idna==3.6
ipykernel==6.29.3
ipython==8.22.2
jedi==0.19.1
Jinja2==3.1.3
joblib==1.3.2
jupyter_client==8.6.1
jupyter_core==5.7.2
kiwisolver==1.4.5
langcodes==3.3.0
MarkupSafe==2.1.5
matplotlib==3.8.3
matplotlib-inline==0.1.6
murmurhash==1.0.10
nest-asyncio==1.6.0
nltk==3.8.1
numpy==1.26.4
openpyxl==3.1.2
packaging==24.0
pandas==2.2.1
parso==0.8.3
pefile==2023.2.7
pillow==10.2.0
platformdirs==4.2.0
preshed==3.0.9
prompt-toolkit==3.0.43
psutil==5.9.8
pure-eval==0.2.2
pybind11==2.12.0
pycparser==2.22
pydantic==2.6.4
pydantic_core==2.16.3
Pygments==2.17.2
pyinstaller==5.13.2
pyinstaller-hooks-contrib==2024.4
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
pywin32==306
pywin32-ctypes==0.2.2
pyzmq==25.1.2
regex==2023.12.25
requests==2.31.0
scikit-learn==1.4.1.post1
scipy==1.12.0
setuptools==69.2.0
six==1.16.0
smart-open==6.4.0
spacy-legacy==3.0.12
spacy-loggers==1.0.5
spacytextblob==4.0.0
srsly==2.4.8
stack-data==0.6.3
textblob==0.15.3
thinc==8.2.3
threadpoolctl==3.3.0
tornado==6.4
tqdm==4.66.2
traitlets==5.14.2
typer==0.9.0
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
wasabi==1.1.2
wcwidth==0.2.13
weasel==0.3.4
whichcraft==0.6.1
wordcloud==1.9.3
zope.event==5.0
zope.interface==6.3
